rename: move JwtUtil to model
