+++ src/main/java/gift/model/JwtUtil.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.model;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.Claims;
